@@ -7,14 +7,15 @@
     "gcb": "./build/src/cli.js",
     "gcbuild": "./build/src/cli.js"
   },
-  "main": "./build/src/index.js",
+  "type": "module",
+  "exports": "./build/src/index.js",
   "types": "./build/src/index.d.ts",
   "files": [
     "build/src",
     "src/.gcloudignore"
   ],
   "engines": {
-    "node": ">=10"
+    "node": ">=12"
   },
   "scripts": {
     "test": "c8 mocha build/test",
@@ -34,30 +35,31 @@
   "author": "Justin Beckwith <justin.beckwith@gmail.com>",
   "license": "MIT",
   "dependencies": {
-    "chalk": "^4.0.0",
-    "globby": "^11.0.0",
+    "chalk": "^4.1.2",
+    "globby": "^12.0.2",
     "googleapis": "^91.0.0",
-    "js-yaml": "^4.0.0",
-    "meow": "^9.0.0",
-    "ora": "^5.0.0",
-    "tar": "^6.0.0",
-    "update-notifier": "^5.0.0"
+    "js-yaml": "^4.1.0",
+    "meow": "^10.1.1",
+    "ora": "^6.0.1",
+    "tar": "^6.1.11",
+    "update-notifier": "^5.1.0"
   },
   "devDependencies": {
-    "@types/js-yaml": "^4.0.0",
+    "@types/js-yaml": "^4.0.4",
     "@types/meow": "^5.0.0",
     "@types/mocha": "^9.0.0",
-    "@types/node": "^16.0.0",
-    "@types/proxyquire": "^1.3.28",
-    "@types/tar": "^4.0.3",
-    "@types/update-notifier": "^5.0.0",
-    "c8": "^7.0.0",
-    "gts": "^3.0.0",
-    "mocha": "^9.0.0",
-    "nock": "^13.0.0",
-    "proxyquire": "^2.1.3",
+    "@types/node": "^16.11.7",
+    "@types/sinon": "^10.0.6",
+    "@types/tar": "^6.1.0",
+    "@types/update-notifier": "^5.1.0",
+    "c8": "^7.10.0",
+    "gaxios": "^4.3.2",
+    "gts": "^3.1.0",
+    "mocha": "^9.1.3",
+    "nock": "^13.2.1",
     "semantic-release": "^18.0.0",
-    "typescript": "~4.3.0"
+    "sinon": "^12.0.1",
+    "typescript": "~4.4.4"
   },
   "c8": {
     "exclude": [
